refactor: migrate to typescript + bump deps
