--- jsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "#/utils/*": [
-        "utils/*"
-      ],
-      "#/components/*": [
-        "components/*"
-      ],
-      "#/styles/*": [
-        "styles/*"
-      ]
-    }
-  }
-}
