+++ package.json
@@ -8,15 +8,25 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@fontsource/rubik": "^4.5.0",
-    "@nivo/calendar": "^0.74.0",
-    "@nivo/line": "^0.74.0",
-    "next": "12.0.7",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "@fontsource/rubik": "^4.5.11",
+    "@nivo/calendar": "^0.80.0",
+    "@nivo/line": "^0.80.0",
+    "next": "13.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "eslint": "8.4.0",
-    "eslint-config-next": "12.0.7"
+    "@nivo/core": "^0.80.0",
+    "@nivo/tooltip": "^0.80.0",
+    "@types/react": "18.0.25",
+    "eslint": "8.27.0",
+    "eslint-config-next": "13.0.2",
+    "prop-types": "^15.8.1",
+    "typescript": "^4.8.4"
+  },
+  "peerDependencies": {
+    "@nivo/core": "^0.80.0",
+    "@nivo/tooltip": "^0.80.0",
+    "prop-types": "^15.8.1"
   }
 }
