+++ pages/index.tsx
@@ -1,12 +1,12 @@
-import styles from "#/styles/total.module.css";
-import { get, set } from "#/utils/cache";
-import { BASE_URL } from "#/utils/constant";
+import styles from "~/styles/total.module.css";
+import { get, set } from "~/utils/cache";
+import { BASE_URL } from "~/utils/constant";
 
 export default function Total({ data }) {
   return (
     <div className={styles.box}>
       <span className={styles.notice}>
-        Total messages since 5th of December 2021
+        Total messages since 6th of November 2022
       </span>
       <span className={styles.total}>{data}</span>
     </div>
@@ -20,7 +20,7 @@ export async function getServerSideProps() {
   }
 
   const res = await fetch(`${BASE_URL}/total`);
-  const total = await res.json();
+  const total = await res.text();
 
   set("total", total);
