--- utils/constant.js
@@ -1,1 +1,0 @@
-export const BASE_URL = "https://teknologi-umum-captcha.fly.dev";
