+++ utils/constant.ts
@@ -1,0 +1,1 @@
+export const BASE_URL = "https://captcha.teknologiumum.com";
